created apiClientWeather
